git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/tools/vislcg3/trunk@1564 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/GrammarParser_parseAddMapReplaceAppend.cpp
@@ -76,7 +76,7 @@ int GrammarParser::parseAddMapReplaceAppend(const UChar *line, uint32_t key) {
 	result->addRule(rule);
 
 	if (name && name[0] && u_strlen(name)) {
-		result->addAnchor(name, result->rules.size()-1);
+		result->addAnchor(name, (uint32_t)(result->rules.size()-1));
 		rule->setName(name);
 	}
 	else {
