+++ src/GrammarParser_parseSelectRemoveIffDelimitMatch.cpp
@@ -74,7 +74,7 @@ int GrammarParser::parseSelectRemoveIffDelimitMatch(const UChar *line, uint32_t 
 	result->addRule(rule);
 
 	if (name && name[0] && u_strlen(name)) {
-		result->addAnchor(name, result->rules.size()-1);
+		result->addAnchor(name, (uint32_t)(result->rules.size()-1));
 		rule->setName(name);
 	}
 	else {
